fixed text color of items

+++ src/components/main/add-new-product/AddNewProduct.jsx
@@ -51,13 +51,12 @@ const AddNewProduct = () => {
         ) {
             addNewProduct(
                 name,
-                parseInt(price),
+                parseFloat(price),
                 type,
                 color,
                 imageAuthor,
                 image
             );
-
             setName("");
             setPrice("");
             setType("");
